migrated schema to its own file
