+++ db/seed.js
@@ -1,11 +1,13 @@
+import * as schema from './schema';
+
 export const addTables = db => {
     db.transaction( tx => {
         tx.executeSql(deleteTable('monsters'))
         tx.executeSql(deleteTable('sessions'))
         tx.executeSql(deleteTable('active_monsters'))
-        tx.executeSql(addMonstersTable);
-        tx.executeSql(addSessionsTable);
-        tx.executeSql(addActiveMonstersTable);
+        tx.executeSql(schema.addMonstersTable);
+        tx.executeSql(schema.addSessionsTable);
+        tx.executeSql(schema.addActiveMonstersTable);
     },transactionError,transactionSuccess)
 };
 
@@ -21,40 +23,6 @@ const transactionSuccess = () => console.log("SUCCESS");
 
 const deleteTable = (name) => `
     DROP TABLE ${name};
-`
-const addMonstersTable = `
-    CREATE TABLE IF NOT EXISTS monsters (
-        id INTEGER PRIMARY KEY,
-        name TEXT,
-        health INTEGER Default 10,
-        armor_class INTEGER DEFAULT 10,
-        strength INTEGER DEFAULT 10,
-        dexterity INTEGER DEFAULT 10,
-        constitution INTEGER DEFAULT 10,
-        intelligence INTEGER DEFAULT 10,
-        wisdom INTEGER DEFAULT 10,
-        charisma INTEGER DEFAULT 10,
-        auto_roll BOOLEAN DEFAULT true,
-        notes TEXT
-    )
-`;
-
-const addActiveMonstersTable = `
-        CREATE TABLE IF NOT EXISTS active_monsters (
-            id INTEGER PRIMARY KEY,
-            monster_id INTEGER NOT NULL,
-            session_id INTEGER NOT NULL,
-            FOREIGN KEY(monster_id) REFERENCES monsters(id),
-            FOREIGN KEY(session_id) REFERENCES sessions(id)
-        )
-`;
-
-const addSessionsTable = `
-   CREATE TABLE IF NOT EXISTS sessions (
-       id INTEGER PRIMARY KEY,
-       timestamp DATETIME DEFAULT CURRENT_TIMESTAMP,
-       active BOOLEAN DEFAULT true
-   )         
 `
 
 const addTestMonster = (tx,name) => {
